requirements.txt
streamlit==1.31.0
groq==0.4.2
python-dotenv==1.0.0
pillow==10.2.0
pytesseract==0.3.10
easyocr==1.7.1
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.2
sentence-transformers==2.3.1
plotly==5.18.0
opencv-python==4.9.0.80
torch==2.1.2
google-generativeai==0.3.2
